# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Hotel Booking API Service Documentation",
        "version": "3.0.0",
        "title": "API Documentation"
    },
    "schemes": [
        "http"
    ],
    "host": "43.204.147.178:3000",
    "basePath": "/",
    "paths": {
...
        "/api/bookinginfo/booking": {
            "post": {
                "summary": "Create a new booking.",
                "description": "Create a new booking with or without guest details.",
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "booking",
                        "description": "Booking information",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "room_id": {
                                    "type": "number",
                                    "description": "Identifier of the booked room.",
                                    "required": true
                                },
                                "guest_id": {
                                    "type": "number",
                                    "description": "Identifier of the guest. Set to null if guest details are provided.",
                                    "required": false
                                },
                                "check_in_date": {
                                    "type": "string",
                                    "format": "date",
                                    "description": "Check-in date for the booking.",
                                    "required": true
                                },
                                "check_out_date": {
                                    "type": "string",
                                    "format": "date",
                                    "description": "Check-out date for the booking.",
                                    "required": true
                                },
                                "booking_date": {
                                    "type": "string",
                                    "format": "date",
                                    "description": "Date and time when the booking was made.",
                                    "required": true
                                },
                                "adults": {
                                    "type": "number",
                                    "description": "Number of adult guests.",
                                    "required": true
                                },
                                "childrens": {
                                    "type": "number",
                                    "description": "Number of child guests.",
                                    "required": true
                                },
                                "total_price": {
                                    "type": "number",
                                    "description": "Total price for the booking.",
                                    "required": true
                                },
                                "payment_status": {
                                    "type": "string",
                                    "enum": [
                                        "paid",
                                        "pending"
                                    ],
                                    "description": "Payment status for the booking. Must be 'paid' or 'pending'.",
                                    "required": true
                                },
                                "created_time": {
                                    "type": "string",
                                    "format": "date-time",
                                    "description": "Timestamp indicating when the booking record was created.",
                                    "required": false
                                },
                                "active": {
                                    "type": "number",
                                    "enum": [
                                        0,
                                        1
                                    ],
                                    "description": "Flag indicating the booking's active status (0 for inactive, 1 for active).",
                                    "required": false
                                },
                                "guest_details": {
                                    "type": "object",
                                    "properties": {
                                        "first_name": {
                                            "type": "string",
                                            "description": "First name of the guest."
                                        },
                                        "last_name": {
                                            "type": "string",
                                            "description": "Last name of the guest."
                                        },
                                        "phone_number": {
                                            "type": "number",
                                            "description": "Phone number of the guest."
                                        },
                                        "email_id": {
                                            "type": "string",
                                            "description": "Email ID of the guest."
                                        },
                                        "date_of_birth": {
                                            "type": "string",
                                            "format": "date",
                                            "description": "Date of birth of the guest."
                                        },
                                        "address": {
                                            "type": "string",
                                            "description": "Address of the guest."
                                        },
                                        "city": {
                                            "type": "string",
                                            "description": "City of residence of the guest."
                                        },
                                        "state": {
                                            "type": "string",
                                            "description": "State of residence of the guest."
                                        },
                                        "country": {
                                            "type": "string",
                                            "description": "Country of residence of the guest."
                                        },
                                        "postal_code": {
                                            "type": "number",
                                            "description": "Postal code of the guest's address."
                                        },
                                        "passport_number": {
                                            "type": "string",
                                            "description": "Passport number of the guest."
                                        },
                                        "address_proof": {
                                            "type": "string",
                                            "description": "Address proof of the guest."
                                        }
                                    },
                                    "required": [
                                        "first_name",
                                        "phone_number",
                                        "email_id",
                                        "date_of_birth",
                                        "address",
                                        "city",
                                        "state",
                                        "country",
                                        "postal_code"
                                    ],
                                    "if": {
                                        "properties": {
                                            "guest_id": {
                                                "type": "number",
                                                "enum": [
                                                    null
                                                ]
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Booking created successfully."
                    },
                    "400": {
                        "description": "Bad request. Invalid input data."
                    }
                }
            },
            "put": {
                "summary": "Update booking information",
                "description": "Update booking information including check-in date, check-out date, booking date, number of adults, number of children, total price, payment status, and updated time.",
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "booking",
                        "description": "Updated booking information",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "id": {
                                    "type": "number",
                                    "description": "The ID of the booking to be updated."
                                },
                                "room_id": {
                                    "type": "number",
                                    "description": "The ID of the room to be updated."
                                },
                                "check_in_date": {
                                    "type": "string",
                                    "format": "date",
                                    "description": "New check-in date for the booking."
                                },
                                "check_out_date": {
                                    "type": "string",
                                    "format": "date",
                                    "description": "New check-out date for the booking."
                                },
                                "adults": {
                                    "type": "number",
                                    "description": "New number of adult guests."
                                },
                                "childrens": {
                                    "type": "number",
                                    "description": "New number of child guests."
                                },
                                "total_price": {
                                    "type": "number",
                                    "description": "New total price for the booking."
                                },
                                "payment_status": {
                                    "type": "string",
                                    "enum": [
                                        "paid",
                                        "pending"
                                    ],
                                    "description": "New payment status for the booking. Must be 'paid' or 'pending'."
                                },
                                "updated_time": {
                                    "type": "string",
                                    "format": "date-time",
                                    "description": "New timestamp indicating when the booking information was updated."
                                }
                            },
                            "required": [
                                "id",
                                "room_id",
                                "check_in_date",
                                "check_out_date",
                                "adults",
                                "childrens",
                                "total_price",
                                "payment_status",
                                "updated_time"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Booking information updated successfully."
                    },
                    "400": {
                        "description": "Bad request. Invalid input data."
                    }
                }
            },
            "delete": {
                "summary": "Cancel a booking",
                "description": "Cancel a booking by providing its ID and a cancellation reason.",
                "parameters": [
                    {
                        "in": "query",
                        "name": "id",
                        "description": "The ID of the booking to be deleted.",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "in": "query",
                        "name": "cancellation_reason",
                        "description": "The reason for cancellation of the booking.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Booking Cancelled successfully."
                    },
                    "400": {
                        "description": "Bad request. Invalid input data."
                    },
                    "404": {
                        "description": "Booking not found."
                    }
                }
            }
        },
        "/api/bookinginfo/booking/{hotelid}": {
            "get": {
                "summary": "Fetch bookings by hotel ID",
                "description": "Retrieve a list of bookings for a specific hotel by providing the hotel's ID.",
                "parameters": [
                    {
                        "name": "hotelid",
                        "in": "path",
                        "description": "The ID of the hotel to retrieve bookings for.",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Successfully received the list of bookings for the specified hotel.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    },
                                    "hotel_id": {
                                        "type": "integer"
                                    },
                                    "room_id": {
                                        "type": "integer"
                                    },
                                    "check_in_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "check_out_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "booking_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "adults": {
                                        "type": "number"
                                    },
                                    "childrens": {
                                        "type": "number"
                                    },
                                    "total_price": {
                                        "type": "number"
                                    },
                                    "payment_status": {
                                        "type": "string",
                                        "enum": [
                                            "paid",
                                            "pending"
                                        ]
                                    },
                                    "cancellation_reason": {
                                        "type": "string"
                                    },
                                    "created_time": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "updated_time": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "active": {
                                        "type": "integer"
                                    },
                                    "guest_details": {
                                        "type": "object",
                                        "properties": {
                                            "first_name": {
                                                "type": "string"
                                            },
                                            "last_name": {
                                                "type": "string"
                                            },
                                            "phone_number": {
                                                "type": "number"
                                            },
                                            "email_id": {
                                                "type": "string"
                                            },
                                            "date_of_birth": {
                                                "type": "string",
                                                "format": "date"
                                            },
                                            "address": {
                                                "type": "string"
                                            },
                                            "city": {
                                                "type": "string"
                                            },
                                            "state": {
                                                "type": "string"
                                            },
                                            "country": {
                                                "type": "string"
                                            },
                                            "postal_code": {
                                                "type": "number"
                                            },
                                            "passport_number": {
                                                "type": "string"
                                            },
                                            "address_proof": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Hotel not found or no bookings available for the specified hotel."
                    }
                }
            }
        }
    }
}